fix: online payment after having disscount

+++ Clothing_Store_POS/Pages/Home/HomePage.xaml.cs
@@ -578,7 +578,7 @@ namespace Clothing_Store_POS.Pages.Home
         private void OnlinePay_Click(object sender, RoutedEventArgs e)
         { 
             
-            string queryString = VNPayHelper.CreatePaymentUrl(TotalAmount);
+            string queryString = VNPayHelper.CreatePaymentUrl((int)TotalAmount);
             Process.Start(new ProcessStartInfo
             {
                 FileName = queryString,
